+++ src/xcp/daq/daq_event.rs
@@ -299,6 +299,14 @@ macro_rules! daq_register {
             $daq_event.add_stack(stringify!($id), &$id as *const _ as *const u8, $id.get_type(), 1, 1, mc_support_data);
         });
     }};
+    // name, event, comment
+    ( $id:ident, $daq_event:expr, $comment:expr ) => {{
+        static ONCE: std::sync::Once = std::sync::Once::new();
+        ONCE.call_once(|| {
+            let mc_support_data = McSupportData::new(McObjectType::Measurement).set_comment($comment);
+            $daq_event.add_stack(stringify!($id), &$id as *const _ as *const u8, $id.get_type(), 1, 1, mc_support_data);
+        });
+    }};
     // name, event
     ( $id:ident, $daq_event:expr ) => {{
         static ONCE: std::sync::Once = std::sync::Once::new();
@@ -323,6 +331,24 @@ macro_rules! daq_register_struct {
             $crate::registry::McRegisterType::mc_register(&$id, $crate::registry::McRegisterTarget::Event($daq_event.get_event_id()), None);
             // Create an instance of the typedef with event relative addressing on stack
             let mc_support_data = $crate::registry::McSupportData::new($crate::registry::McObjectType::Measurement);
+            $daq_event.add_stack(
+                stringify!($id),
+                &$id as *const _ as *const u8,
+                $crate::registry::McValueType::new_typedef($crate::registry::McRegisterType::mc_type_name_value(&$id)),
+                1,
+                1,
+                mc_support_data,
+            );
+        });
+    }};
+    // Name, event, comment
+    ( $id:ident, $daq_event:expr, $comment:expr ) => {{
+        static ONCE: std::sync::Once = std::sync::Once::new();
+        ONCE.call_once(|| {
+            // Register the typedef for the struct (no instance), event-relative addressing
+            $crate::registry::McRegisterType::mc_register(&$id, $crate::registry::McRegisterTarget::Event($daq_event.get_event_id()), None);
+            // Create an instance of the typedef with event relative addressing on stack
+            let mc_support_data = $crate::registry::McSupportData::new($crate::registry::McObjectType::Measurement).set_comment($comment);
             $daq_event.add_stack(
                 stringify!($id),
                 &$id as *const _ as *const u8,